+++ packages/rubocop-powerhome/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Adds Migration/RenameColumn and Migration/RenameTable cops [#380](https://github.com/powerhome/power-tools/pull/380)
 - Standardize all libs to support ruby 3.0, ruby 3.3 x rails 6.1 through rails 7.2 [#359](https://github.com/powerhome/power-tools/pull/359)
 
 ## [0.5.6] - 2025-09-09